+++ src/firebase/auth.js
@@ -1,25 +1,6 @@
 import {
-  auth, provider, createUserWithEmailAndPassword,
-  signInWithEmailAndPassword, onAuthStateChanged, signOut, signInWithPopup,
-} from './configuraciones.js';
-
-export const createUser = (email, password) => {
-  return createUserWithEmailAndPassword(auth, email, password)
-      .then((userCredential) => {
-        const user = userCredential.user;
-        // console.log('usario creado', {userCredential});
-        console.log('usario creado', {user});
-        console.log(userCredential.user);
-      });
-};
-
-export const signIn = (email, password) => {
-  return signInWithEmailAndPassword(auth, email, password)
-      .then((userCredential) => {
-        const user = userCredential.user;
-        console.log('Inició sesión', {user});
-      });
-};
+  auth,
+  onAuthStateChanged} from './configuraciones.js';
 
 export const onUser = (user) => {
   return onAuthStateChanged(auth, user);
@@ -31,16 +12,4 @@ export const onUser = (user) => {
   //     const uid = null;
   //   }
   // });
-};
-
-export const signOutUser = () => {
-  return signOut(auth);
-};
-
-export const googleAuth = () => {
-  return signInWithPopup(auth, provider).then((result) => {
-    console.log(result);
-  }); /* .catch((error) => {
-console.log(error);
-});  */
 };
