+++ src/firebase/auth_google_signin_popup.js
@@ -1,0 +1,9 @@
+import {auth, provider, signInWithPopup} from './configuraciones.js';
+
+export const googleAuth = () => {
+  return signInWithPopup(auth, provider).then((result) => {
+    console.log(result);
+  }); /* .catch((error) => {
+console.log(error);
+});  */
+};
