+++ src/firebase/auth_sign_out.js
@@ -1,0 +1,5 @@
+import {auth, signOut} from './configuraciones.js';
+
+export const signOutUser = () => {
+  return signOut(auth);
+};
