+++ src/firebase/auth_signup_password.js
@@ -1,0 +1,12 @@
+import {auth, createUserWithEmailAndPassword} from './configuraciones.js';
+
+export const createUser = (email, password) => {
+  return createUserWithEmailAndPassword(auth, email, password).then(
+      (userCredential) => {
+        const user = userCredential.user;
+        // console.log('usario creado', {userCredential});
+        console.log('usario creado', {user});
+        console.log(userCredential.user);
+      },
+  );
+};
