+++ src/main.js
@@ -1,7 +1,12 @@
 // Este es el punto de entrada de tu aplicacion
 
 // import {myFunction} from './lib/index.js';
-import {createUser, signIn, googleAuth} from './firebase/auth.js';
+import {createUser} from './firebase/auth_signup_password.js';
+import {signIn} from './firebase/auth_signin_password.js';
+import {googleAuth} from './firebase/auth_google_signin_popup';
+
+
+// import {signOutUser} from './auth_sign_out.js';
 
 // myFunction();
