+++ src/style.css
@@ -5,7 +5,7 @@
     --color-8:#421987;
     --color-7:#51279B;
     --color-6:#653CAD;
-    --Color-5:#724BB7;
+    --color-5:#724BB7;
     --color-4:#8662C7;
     --color-3:#A081D9;
     --color-2:#CFBCF2;
@@ -147,8 +147,8 @@ html{
 }
 
 .button{
-    border: solid 1px #421987;
-    background-color: var(--Color-5);
+    border: solid 1px var(--color-5);
+    background-color: var(--color-5);
     color: white;
     transition:background-color .5s;
     cursor: pointer;
@@ -158,11 +158,16 @@ html{
 }
 
 .formLogIn>input,.formResgister>input,.button{
-    border:solid var(--color-1);
+    /* border:solid var(--neutral-1); */
     border-radius: 8px;
     height: 36px;
     width:266px;
     margin: 12px;
+}
+
+.formLogIn>input{
+    background-color: var(--neutral-2);
+    border:solid var(--neutal-2);
 }
 
 @media (min-width:450px){
